requirements.txt
blueprint==3.4.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.1.2
PyMySQL==0.9.3
pytesseract==0.3.4
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
